+++ pygromos/simulations/modules/general_analysis_modules.py
@@ -1,0 +1,8 @@
+"""
+    ideas:
+        * Automatic Equilibration detection
+        * property analysis
+        * default analysis for emin ? eq?
+
+    Start with local submission!
+"""
