+++ pygromos/tests/test_simulation_blocks/test_simulation_runner_blocks.py
@@ -1,6 +1,8 @@
-import unittest, os, tempfile, numpy as np
+import unittest, numpy as np
 
-from pygromos.simulation_runner import simulation_building_blocks
+import pygromos.simulations.modules.general_simulation_modules
+import pygromos.simulations.modules.ti_modules
+from pygromos.simulations.modules import preset_simulation_modules
 from pygromos.hpc_queuing.submission_systems.Submission_Systems import DUMMY
 
 from pygromos.tests.in_testfiles import in_test_file_path
@@ -10,7 +12,7 @@ from pygromos.tests.test_simulation_blocks import out_test_root_dir
 
 class test_simulation_blocks(unittest.TestCase):
     submissionSystem = DUMMY()
-    sim_block = simulation_building_blocks.emin
+    sim_block = preset_simulation_modules.emin
     input_cnf_path = in_test_file_path+"/small_system/6J29.cnf"
     input_top_path = in_test_file_path+"/small_system/6J29.top"
 
@@ -24,28 +26,28 @@ class test_simulation_blocks(unittest.TestCase):
     def test_smulation(self):
         from pygromos.data.simulation_parameters_templates import template_md
         self.gromSystem.imd = template_md
-        simulation_building_blocks.simulation(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                      submission_system=self.submissionSystem)
+        pygromos.simulations.modules.general_simulation_modules.simulation(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                                                           submission_system=self.submissionSystem)
 
     def test_emin(self):
-        simulation_building_blocks.emin(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                      submission_system=self.submissionSystem)
+        preset_simulation_modules.emin(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                       submission_system=self.submissionSystem)
 
     def test_sd(self):
-        simulation_building_blocks.sd(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                      submission_system=self.submissionSystem)
+        preset_simulation_modules.sd(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                     submission_system=self.submissionSystem)
 
     def test_md(self):
-        simulation_building_blocks.md(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                      submission_system=self.submissionSystem)
+        preset_simulation_modules.md(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                     submission_system=self.submissionSystem)
 
     def test_lam_window(self):
-        simulation_building_blocks._TI_lam_step(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                      submission_system=self.submissionSystem)
+        pygromos.simulations.modules.ti_modules._TI_lam_step(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                                             submission_system=self.submissionSystem)
 
     def test_ti_sampling(self):
         from pygromos.data.simulation_parameters_templates import template_md
         self.gromSystem.imd = template_md
-        simulation_building_blocks.TI_sampling(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
-                    lambda_values = np.arange(0, 1.1, 0.1), subSystem = self.submissionSystem,
-                    n_simulation_repetitions = 3, n_equilibrations = 1)
+        pygromos.simulations.modules.ti_modules.TI_sampling(in_gromos_system=self.gromSystem, project_dir=self.tmp_test_dir,
+                                                            lambda_values = np.arange(0, 1.1, 0.1), subSystem = self.submissionSystem,
+                                                            n_simulation_repetitions = 3, n_equilibrations = 1)
